Refactor - sessions endpoints to separate module
